Rename clip_extractors to clip
